refactor: Separação das páginas em pastas próprias
