--- src/style/home.module.css
@@ -1,86 +1,0 @@
-* {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-}
-
-body {
-    width: 100%;
-    min-height: 100vh;
-    display: flex;
-    padding: 0px 111px;
-    justify-content: center;
-    align-items: center;
-    background: #000;
-}
-
-.container {
-    color: red;
-}
-
-.titulo {
-    display: flex;
-    /* width: 698px;
-    height: 96px; */
-    justify-content: center;
-    align-items: center;
-    flex-shrink: 0;
-    border-radius: 2px;
-    border: 1px solid rgba(0, 0, 0, 0.10);
-}
-
-h1 {
-    flex-direction: column;
-    color: #FFF;
-    text-align: center;
-    font-family: Spinnaker;
-    font-size: 64px;
-    font-style: normal;
-    font-weight: 400;
-    line-height: normal;
-    padding: 15px;
-}
-
-/* .linha {
-    width: 703px;
-    height: 5px;
-} */
-
-.botao {
-    display: flex;
-    /* flex-direction: row;
-    align-content: stretch; */
-    justify-content: center;
-    padding: 15px;
-}
-
-img {
-    padding-bottom: 100px;
-}
-
-/* button {
-    width: 387px;
-    height: 72px;
-    flex-shrink: 0;
-} */
-
-footer {
-    /* display: flex;
-    width: 1440px;
-    padding-bottom: 0px;
-    justify-content: center;
-    align-items: center; */
-    position: fixed;
-    bottom: 0;
-    left: 50;
-    background: #000;
-    color: rgba(255, 255, 255, 0.25);
-    font-family: Spinnaker;
-    font-size: 20px;
-    display: flex;
-    width: 101px;
-    height: 29px;
-    flex-direction: column;
-    justify-content: center;
-
-}
